requirements.txt
crewai==0.50.0
langchain==0.1.16
langchain-community==0.0.38
langchain-openai==0.1.3
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
python-dateutil==2.8.2
pandas==2.1.3
json5==0.9.14
